Using variants of POST for apic vs capic
The commit 858e9b1caab098bb5aea9e55a47dee2ac24a1a30
started sending the POST by assigning the request body
to the json argument. However, this breaks the POST
to the APIC which accepts the dict to be form-encoded
and thus requires setting the data argument.

This is being fixed here by using the relevant variant
of the POST depending on whether the POST is being made
to APIC or CAPIC. Ideally the same variant should work
for both and it should be investigated further.