Changes for openshift-4.4-openstack flavor
Similar to 4.3, just without kube_api_vlan or node bridge domain, and open ports 2380 and 9978 for etcd

A new option in aci_config called "disable_node_bd_creation" which causes node_bd to be created only if the option is set to False(True by default).
This option also takes care of not associating the node EPG to any BD in this case.