Adding RKE flavor
As of now this flavor generates the network provider part of
cluster.yml to be used for RKE.Using 1.1.5-rc0, since that is
the minimum supported RKE version.

TODO:
Will add the Rancher RKE template once Rancher config is operational.
Validating the final generated yaml from RKE side is possible through
Jenkins, but not possible in the UTs.

Signed-off-by: Kiran Shastri <shastrinator@gmail.com>