To make opflex metrics configurable with acc-provision
We introduced a new field named opflex_agent_prometheus to provision_input to enable/disable opflex metrics which is set default to false. We have conditional check of this field for port info and annotations to check whether it is set to True or False in order to add them to opflex-agent container and aci-containers-host DaemonSet respectively.