Removing extra space in versions config file
This does not have any functional impact on acc-provision itself,
however some automation which reads this file is failing on this.
That automation can be made more robust, but also cleaning up here
since the extra space is undesirable.

Also updated OCP 4.12 ESX expected test output which was missed
in earlier merge.