Unit tests - solve cherry-pick merge conflicts
Whenever any parameter is updated/added in aci-operator-config configmap
it generates a new base64 encoded string in  *.operator_cr.kube.yaml
*.kube.yaml and *ConfigMap-aci-operator-config.yaml.
cherry-pick of that commit from master to other branch(e.g kmr2,
cko-mvp-1) results in merge conflict as git is not able to merge
encoded content.

Fix to always store aci-operator-config configmap data in an un-encoded
format in the testdata in above mentioned files and encode it after
loading it in the unit tests just before performing the comparison between
generated and expected output.