Validate openshift system_id
OpenShift accepts as cluster name only a valid DNS Name.
If we create a cluster with say an '_' (underscore) in it openshift installation
will fail. Valid DNS chars are a to z, the digits 0 to 9 and the '-' (hypen).
Considering that we can't use '-' (hyphen) due to ACI limitation we should just
accept a-z and 0-9 and all MUST be lowercase.