Reduce the size of acc-provision input pushed to APIC
We push the input for acc-provision to vmmInjectedClusterInfo.accProvisionInput.
The max length supported by the APIC for the field is 8192.
Some users are hitting this limit with the increased number of comments and
whitespace in the input file.
Strip out comments and whitespace by loading/dumping the yaml file before POST

Signed-off-by: Tom Flynn <tom.flynn@gmail.com>