Create HPP Direct CRDs always
Create the hostprotPol and hostprotRemoteIpContainer CRDs and
roles irrespective of the enable_hpp_direct flag set to true
or not.

This is to allow clearing the CRs when switching from hpp direct
enabled to disabled and also to allow enabling/disabling hpp
direct without provisioning which is useful in test automation.